state: remove testing only method from public API
Move AllWatcher.All() method from the public API, we can make it part of
the exported API in test scope alone.
